util, a package
I added a package based around utilities.
I moved every validation in validation.py.
